Move calico releated tests to a separate file
